added different data types to IPC

+++ src/RLFuzzing.py
@@ -7,7 +7,7 @@ class RLFuzzing:
     def __init__(self):
         return
 
-    def recieve_messages(self, BUFF_SIZE_RECIEVER = 32):
+    def recieve_messages(self, BUFF_SIZE_RECIEVER=32):
         try:
             mq_reciever = sysv_ipc.MessageQueue(1, sysv_ipc.IPC_CREAT)
 
@@ -20,7 +20,7 @@ class RLFuzzing:
                 print(f"mtype: {mtype}")
                 self.send_messenges(mtype)
             elif mtype == 2:
-                afl_fsrv_map_size = np.frombuffer(message, dtype=np.uintc)
+                afl_fsrv_map_size = np.frombuffer(message, dtype=np.uint8)
                 # afl_fsrv_map_size = message.decode()
                 print(f"afl->fsrv.map_size: {afl_fsrv_map_size}")
                 print(f"mtype: {mtype}")
@@ -29,11 +29,11 @@ class RLFuzzing:
         except sysv_ipc.ExistentialError:
             print("ERROR: message queue creation failed")
 
-    def send_messenges(self, mtype, BUFF_SIZE_SENDER = 64):
+    def send_messenges(self, mtype, BUFF_SIZE_SENDER=64):
         if mtype == 1:
             msg_npy = np.arange(BUFF_SIZE_SENDER, dtype=np.double).reshape((2,BUFF_SIZE_SENDER//2))
         elif mtype == 2:
-            msg_npy = np.arange(BUFF_SIZE_SENDER, dtype=np.uintc).reshape((2,BUFF_SIZE_SENDER//2))
+            msg_npy = np.arange(BUFF_SIZE_SENDER, dtype=np.uint8).reshape((2,BUFF_SIZE_SENDER//2))
 
         try:
             mq_sender = sysv_ipc.MessageQueue(2, sysv_ipc.IPC_CREAT)
